requirements.txt
mysql-connector-python
PySide6
bcrypt
pymupdf
numpy